// pos-fe/src/interface/DataTransaction.interface.ts
// export interface TransactionItemDto {

// }

export interface DataTransactionDto {
    Id: string;
    transactionId: string;
    grossAmount: string;
    currency: string;
    orderId: string;
    paymentType : string;
    transactionStatus: string;
    statusMessage: string;
    transactionTime: string;
    petugasId: string;
}